+++ AndroidGarage/data/src/commonMain/kotlin/com/chriscartland/garage/data/ktor/KtorNetworkButtonDataSource.kt
@@ -15,7 +15,7 @@
  *
  */
 
-package com.chriscartland.garage.internet
+package com.chriscartland.garage.data.ktor
 
 import co.touchlab.kermit.Logger
 import com.chriscartland.garage.data.NetworkButtonDataSource
